Name that Country
This is a fun trivia game with two levels. You can choose to type in the name of a country in level or guess the capital city of a country. A flag is displayed, if the user recognizes this (or wishes to guess) they type the name of the flag into the textbox and hits enter clicks submit. If they are correct, they are given a point. If they are incorrect, they are not given a point and are told the correct answer. There is a timer counting down and the user must get as many points as possible before the timer runs out. There is a skip question button. When the timer runs out, the game ends and a stats screen appears with the user's score. The second level is similar, the name of the country and the flag are displayed and the user must guess the capital city.